+++ CHANGES.rst
@@ -1,6 +1,8 @@
 Changes
 =======
 
+- Compiler now reports an error if an expression type is unknown.
+
 - Edge-case issue where an expression result was actually not a
   dynamic value, but a static string (e.g. ``string: Hello``).